+++ src/main/java/com/ewp/crm/models/Mentor.java
@@ -2,7 +2,10 @@ package com.ewp.crm.models;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.ManyToMany;
 import javax.persistence.Table;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Table(name = "mentor")
@@ -12,6 +15,10 @@ public class Mentor extends User {
      */
     @Column(name = "mentor_show_only_my_clients")
     private boolean showOnlyMyClients;
+
+    //Связь с направлениями
+    @ManyToMany(mappedBy = "clients")
+    private List<Course> courses = new ArrayList<>();
 
     public Mentor() {
     }
@@ -27,5 +34,13 @@ public class Mentor extends User {
 
     public void setShowOnlyMyClients(boolean showOnlyMyClients) {
         this.showOnlyMyClients = showOnlyMyClients;
+    }
+
+    public List<Course> getCourses() {
+        return courses;
+    }
+
+    public void setCourses(List<Course> courses) {
+        this.courses = courses;
     }
 }
